avoid decreasing precision for ASE atoms

+++ dpdata/plugins/ase.py
@@ -52,8 +52,8 @@ class ASEStructureFormat(Format):
             "atom_names": atom_names,
             "atom_numbs": atom_numbs,
             "atom_types": atom_types,
-            "cells": np.array([cells]).astype("float32"),
-            "coords": np.array([coords]).astype("float32"),
+            "cells": np.array([cells]),
+            "coords": np.array([coords]),
             "orig": np.zeros(3),
             "nopbc": not np.any(atoms.get_pbc()),
         }
@@ -87,15 +87,15 @@ class ASEStructureFormat(Format):
         forces = atoms.get_forces()
         info_dict = {
             **info_dict,
-            "energies": np.array([energies]).astype("float32"),
-            "forces": np.array([forces]).astype("float32"),
+            "energies": np.array([energies]),
+            "forces": np.array([forces]),
         }
         try:
             stress = atoms.get_stress(False)
         except PropertyNotImplementedError:
             pass
         else:
-            virials = np.array([-atoms.get_volume() * stress]).astype("float32")
+            virials = np.array([-atoms.get_volume() * stress])
             info_dict["virials"] = virials
         return info_dict
